Crown one true king
Now that the settings singleton is not special, we only need one of them
